sql/requirements.txt
mysql-connector-python==8.0.16
protobuf==3.9.0
six==1.12.0
